Create Agama installation for full encryption
