+++ src/pages/login-as-root-page.ts
@@ -1,9 +1,9 @@
 import puppeteer, { type Page } from "puppeteer-core";
 
 export class LoginAsRootPage {
-    readonly page: Page;
-    readonly passwordInput = () => this.page.locator("input#password");
-    readonly logInButton = () => this.page.locator("button[type='submit']");
+    private readonly page: Page;
+    private readonly passwordInput = () => this.page.locator("input#password");
+    private readonly logInButton = () => this.page.locator("button[type='submit']");
 
     constructor(page: Page) {
         this.page = page;
